support the address-data report entry
rfc6352 defines CARDDAV:address-data as a way to include the body of a
card in a report result. it also specifies that the client can restrict
the list of properties that should be returned about the resource in the
CARDDAV:address-data line in the report request, but implementing the
simple case (as was done with calendar-data before) is sufficient to
make the davdroid client happy.